requirements.txt
pytest
psycopg2-binary
python-dotenv
flask
matplotlib
mpld3
